add template for eslintignore
Following the introduction of eslint as a tool for consistent code style,
the neccessity to keep files out of this stylecheck arose. Examples of
excludeable files are transpiled outputs. The added template
provides an enhanceable boilerplate ignorefile. It will replace the now
as ignorefile for eslint used .gitignore (see updated gulp-tasks).