replace prettier with equivalent eslint rules
How noble the intents for the usage of prettier tend to be,
they seem slightly superfluous if you put an eslint documentation
next to it. That means, you can achieve all prettier rules
with eslint, too. This commit introduces this rules replacement.

Furthermore, it widen and introduces rules (up for discussion):
- line width should be maximum 100 signs instead of 80 (because 2018 FullHD)
- spaces inside objects should improve readability of object content